# Small optimizations to trimesh pd_mesh utility