@jbednar 2018-01-16
Looks and sounds good to me, assuming the tests pass!